pyproject.toml
[tool.poetry]
name = "Bucky"
version = "0.1.0"
description = "Matrix bot"
authors = ["Vincent Carlos <git@finzzz.net>"]

[tool.poetry.dependencies]
python = "^3.7"
matrix-nio = {extras = ["e2e"], version = "^0.16.0"}
importlib-metadata = "^3.10.1"
pyinstaller = "^4.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
